Terminado Crud PreciosProducto

+++ src/main/java/com/nanoka/weblibreria/dao/MovimientoProducto.java
@@ -1,9 +1,6 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package com.nanoka.weblibreria.dao;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -17,7 +14,7 @@ public class MovimientoProducto {
     private int id;
     private int productoId;
     private int cantidad;
-    private double monto;
+    private BigDecimal monto;
     private String razon;
     private String fecha;
 }
